magnetic field color map correction2

+++ data analysis/B_map.py
@@ -13,7 +13,7 @@ parser.add_argument('--beta', type=int, help='beta')
 parser.add_argument('--b_high', type=float, help='beta high')
 parser.add_argument('--b_low', type=float, help='beta low')
 parser.add_argument('--rank', type=int, help='rank')
-parser.add_argument('--N_steps', type=float, help='N_steps' )
+parser.add_argument('--N_steps', type=int, help='N_steps' )
 
 # Parse the command-line arguments
 args = parser.parse_args()
